Put userConverter function into user.helper file
