+++ services/user.helper.ts
@@ -1,7 +1,44 @@
 import { FIREBASE_DB } from '@/firebaseConfig';
 import { GetUserByUsernameResponse } from '@/types/social';
-import { collection, getDocs, query, where } from 'firebase/firestore';
-import { userConverter } from './user';
+import { User } from '@/types/user';
+import { Workout } from '@/types/workout';
+import { fromTimestampToDate } from '@/utils/general';
+import {
+  collection,
+  getDocs,
+  query,
+  QueryDocumentSnapshot,
+  Timestamp,
+  where,
+} from 'firebase/firestore';
+
+export const userConverter = {
+  toFirestore: (data: User) => data,
+  fromFirestore: (snap: QueryDocumentSnapshot) => {
+    // Convert workoutHistory startedAt to Date object
+    const data = snap.data() as Omit<User, 'createdAt' | 'workoutHistory'> & {
+      createdAt: Timestamp;
+      workoutHistory: Array<
+        Omit<Workout, 'startedAt'> & {
+          startedAt: Timestamp;
+        }
+      >;
+    };
+
+    const newWorkoutHistory = data.workoutHistory.map((workout) => {
+      return {
+        ...workout,
+        startedAt: fromTimestampToDate(workout.startedAt),
+      };
+    });
+
+    return {
+      ...data,
+      createdAt: fromTimestampToDate(data.createdAt),
+      workoutHistory: newWorkoutHistory,
+    };
+  },
+};
 
 export const getUserByUsername: (
   username: string,
